fixed a very bad bug

+++ Taglib/NSString_Filename.m
@@ -69,7 +69,7 @@
                     } else {
                         NSMutableString * tmpValue = [NSMutableString stringWithString:frames[@"track"]];
                         if (trackSize > tmpValue.length) {
-                            for (NSUInteger i = trackSize - tmpValue.length; i > 0; i--) {
+                            for (int i = (int)(trackSize - tmpValue.length); i > 0; i--) {
                                 [tmpValue insertString:@"0" atIndex:0];
                             }
                         }
